requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Config & Data Validation
pydantic
pydantic-settings
python-dotenv
typing-extensions

# Logging
structlog

# Observability
prometheus-fastapi-instrumentator

# AI & Vector DB
qdrant-client
sentence-transformers
# torch Dockerfile'da özel olarak kuruluyor

# Web Scraping
beautifulsoup4
requests

# PostgreSQL
psycopg2-binary

# 
playwright